backend/ai_agents/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-ai==0.1.0
langchain==0.1.9
langchain-openai==0.0.5
langgraph==0.1.11
sqlalchemy==2.0.27
python-dotenv==1.0.0
websockets==12.0
python-multipart==0.0.9